requirements.txt
langchain>=0.3.0,<0.4.0
langchain-google-genai>=2.0.0,<3.0.0
langgraph>=0.2.0,<0.3.0
toolbox-langchain
google-generativeai>=0.8.0,<1.0.0
pydantic>=2.0.0,<3.0.0